% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_processing.R
\name{cor_processing}
\alias{cor_processing}
\title{Correlation processing}
\usage{
cor_processing(data)
}
\arguments{
\item{data}{\itemize{
\item data frame with gene length and size per organism
}}
}
\value{
correlate - dataframe with calculations
}
\description{
This function performs correlation calculations
}
