% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_processing.R
\name{cor_processing_top_5}
\alias{cor_processing_top_5}
\title{Correlation processing top 5}
\usage{
cor_processing_top_5(data)
}
\arguments{
\item{data}{\itemize{
\item data frame with gene length and size per organism
}}
}
\value{
gpo_length_size_top_5 - dataframe with calculations
}
\description{
This function performs correlation calculations to find the 5 strongest relationships
and groups by organism and add a unique sort id
}
