% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_processing.R
\name{cor_scatterplot}
\alias{cor_scatterplot}
\title{Correlation calc display}
\usage{
cor_scatterplot(data)
}
\arguments{
\item{data}{\itemize{
\item data frame with gene length and size per organism
}}
}
\description{
This function calculates and displays Pearson correlation coefficients
}
