% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_length_size_calc.R
\name{gene_length_size_calc}
\alias{gene_length_size_calc}
\title{Gene length size calculation}
\usage{
gene_length_size_calc(organism, genes)
}
\arguments{
\item{organism}{\itemize{
\item dataframe with organism data
}}

\item{genes}{\itemize{
\item data with gene data
}}
}
\value{
gpo_length_size - a dataframe with gene lengths and sizes
}
\description{
This function calculates the gene length and size and
selects the columns needed for further processing
}
