% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpo_processing.R
\name{gpo_output_processing}
\alias{gpo_output_processing}
\title{GPO output}
\usage{
gpo_output_processing(data, group_id_param)
}
\arguments{
\item{data}{\itemize{
\item data frame with stats data
}}

\item{group_id_param}{\itemize{
\item integer parameter with group ID
}}
}
\value{
group_output - filtered dataframe
}
\description{
This function prepares output stats for the GPO stats tables
by filtering the data by group ID
}
