% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpo_processing.R
\name{gpo_processing}
\alias{gpo_processing}
\title{GPO count}
\usage{
gpo_processing(organism, genes)
}
\arguments{
\item{organism}{\itemize{
\item dataframe with organism data
}}

\item{genes}{\itemize{
\item data with gene data
}}
}
\value{
gpo_count - a dataframe with genes per organism counts
}
\description{
Create genes per organism count
}
