% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_count_processing.R
\name{summary_count_processing}
\alias{summary_count_processing}
\title{Data summary Function}
\usage{
summary_count_processing(organism, genes)
}
\arguments{
\item{organism}{\itemize{
\item dataframe with organism data
}}

\item{genes}{\itemize{
\item data with gene data
}}
}
\value{
summary_count_df - a dataframe with category counts
}
\description{
This function creates the a dataframe with the counts in various categories
}
